+++ docs/run-ra-docs.sh
@@ -1,0 +1,3 @@
+docker rm -f metis-ra-docs
+docker build -t ra-docs:latest .
+docker run -d -p  4405:80 --name=metis-ra-docs ra-docs