+++ ra-back/run-ra-back.sh
@@ -1,0 +1,3 @@
+docker rm -f metis-ra-back
+docker build -t ra-back:latest .
+docker run -d -p  4401:8080 --name=metis-ra-back ra-back