+++ changelog.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ## [1.10.3] - 2020-07-08
+### Added
+- Docker run script files
+### Changed
+- all addr change to localhost (without MongoDb on Tebe)
+
+## [1.10.3] - 2020-07-08
 ### Changed
 - Documentation: info about stacks, users, technologies, etc.
 - Up version to 1.10.3